refactor(config): delete the config.defaults re-export aggregator
config/defaults.py existed only to re-export five default classes from their
real owners. Four of the five had no users at all; the fifth, OutputDefaults,
had three importers. Repoint those at aiperf.config.artifacts and delete the
module so each default class has exactly one import path.
