--- src/aiperf/config/defaults.py
@@ -1,9 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-"""Aggregator re-exports for config default classes."""
-
-from aiperf.config.artifacts import OutputDefaults as OutputDefaults
-from aiperf.config.dataset.defaults import InputDefaults as InputDefaults
-from aiperf.config.dataset.defaults import InputTokensDefaults as InputTokensDefaults
-from aiperf.config.endpoint import EndpointDefaults as EndpointDefaults
-from aiperf.config.runtime import ServiceDefaults as ServiceDefaults
